feat: Add initial draft for training, modify directory and file structure
