+++ cfg/cfg.yaml
@@ -18,3 +18,8 @@ test_ratio: 0.1
 
 # Random state for splitting the dataset
 random_state: 42
+
+# Model trining parameters
+training:
+  num_epochs: 100
+  learning_rate: 1e-2
