+++ src/models/parabolic_motion_model.py
@@ -1,0 +1,14 @@
+import torch.nn as nn
+
+class ParabolicMotionModel(nn.Module):
+    def __init__(self, input_size, hidden_size, output_size):
+        super(ParabolicMotionModel, self).__init__()
+        self.fc1 = nn.Linear(input_size, hidden_size)
+        self.relu = nn.ReLU()
+        self.fc2 = nn.Linear(hidden_size, output_size)
+
+    def forward(self, x):
+        out = self.fc1(x)
+        out = self.relu(out)
+        out = self.fc2(out)
+        return out
